requirements.txt
diffusers==0.21.4
openai==0.28.1
Pillow==9.3.0
Pillow==10.1.0
streamlit==1.28.0
torch==2.1.0+cu118
